Move some actions from RoomListView to Room
Rooms are now deleted by using the standard "destroy" method of Backbone
models instead of calling the endpoint directly from the view. In a
similar way, the current user now removes herself from the room also by
destroying the model, but with a special URL. This will make possible to
listen to and handle the "destroy" event of the model from other places.

Leaving the current room when it is destroyed or removed was also moved
to the model; for consistency, the model now provides a method to join
the room too (which, in the end, will also synchronize the model).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>