Fix Chat tab not shown in Files app if Sharing tab was not opened
When the details view is opened for a file the "shareTypes" of the
FileInfo are initially strings; they are converted to integers only once
the "Sharing" tab has been opened. Thus, now the "shareTypes" are always
"casted" to integers before performing a strict comparison against the
"SHARE_TYPE_XXX" constants.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>