Set room left/deleted message as a response to events
The "Connection" object should not change the UI directly, so now the
empty content message when the current room is left or deleted it is set
as a response to the "leaveCurrentRoom" and "destroy" events.

As the "deleter" parameter of "leaveCurrentRoom" was only used to show
one message or the other it is no longer needed, so it was removed.

The empty content message set when the room is deleted was also updated
to match the initial empty content message.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>