Delay reload of the message list
When "setFileInfo" is called the tab is not visible yet, so reloading
the message list does nothing; the reload needs to be delayed until the
next event loop to ensure that the tab will be visible.

Even without this fix the list was being shown as expected due to
joining the room causing the messages to be fetched and thus the list
being reloaded when received. However, joining to the room before
showing the tab would reveal the problem (as will happen once the call
button is added).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>