Prevent breaking the call when the Files sidebar is closed in Chromium
Chromium seems to drop a stream when the element it is attached to is
detached or reparented. The sidebar in the Files app is open and closed
using a jQuery animation, which reparents the whole sidebar and then
restores it at the end of the animation, so closing the sidebar breaks
an ongoing call in Chromium. To prevent that, during a call the
functions to open and close the sidebar are replaced with custom
versions that do not use an animation.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>