Set room icons in the template
"onRender" is always called after the template has been rendered, so
there is no need to remove the icons for other types of rooms; even if
the room type has changed the element will not use have any of the
previous CSS classes, it will be a new element as defined in the
template. Due to this the room icons can be set in the template itself
instead of in "onRender".

Avatars, on the other hand, are set using JavaScript, so they still need
to be set in "onRender"; the name of the called function has been
changed to better reflect is current purpose.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>