Show working icon in call button while joining or leaving a call
To provide a stronger feedback that the action is still going on now a
working/loading icon is shown in the call button while joining or
leaving a call.

In the Files app the call button width depends on its text (in the
regular Talk UI it depends on the sidebar width), so the padding was
increased to always make room for the icon, even when hidden, to prevent
a sudden increase of the width when the icon is shown.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>