Add full Handlebars library
In Nextcloud 16 the Handlebars library shipped in the server is no
longer the full Handlebars library, but a trimmed down version that does
not support compiling Handlebars templates at runtime.

Currently Talk still needs to compile Handlebars templates at runtime,
so for the time being the full Handlebars library that was included
before in the server is now included in Talk.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>