Extend "has direct access to file" check also to files in shared folders
Before a file was seen as directly accessible by the user if the user
received the file through a user, group, circle or room share. Now files
that are a descendant of a folder that meets those conditions are seen
as directly accessible by the user too.

Due to this now the rooms for files in a shared folder can be accessed
by any user that has access to the folder; it is no longer needed to
explicitly share the file too with them.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>