Query the server if it is unknown if Talk sidebar is supported for file
In some cases it is not possible to know if the Talk sidebar is
supported for a file or not just from the data in the FileInfo. For
example, the Talk sidebar is supported for files in a folder shared by
the current user, but the FileInfo does not provide any hint about that
case (it does, on the other hand, when the file is in a folder shared by
another user with the current one).

In those cases it is necessary to query the server to know whether the
room for the file is accessible by the user or not. Once that is known
everything works as before; if the room is accessible then the Talk
sidebar is supported for the file and the chat is shown; if it is not
accessible then the message hinting the user to share the file is shown
instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>