Do not show call button while the "UI not ready" placeholder is shown
When the "UI not ready" placeholder is shown the other elements in the
tab are hidden using a CSS trick to hide all its next siblings. However,
the call button was prepended to the tab, so it was a previous sibling
of the placeholder and thus it was not hidden. Now it is inserted based
on the chat view instead to ensure that it will be a next sibling of the
placeholder.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>