Fix multiple mentions which are used directly behind each others
The editable div inserts non-breaking spaces (url encode %C2%A0) after a mention.
When two mentions are right behind each others, the later one is not matched by the
regex later, because the lookahead on the @ finds only non-breaking content.
Manually adding a space in between the two mentions solves this already.

But instead we replace them with normal spaces before sending them to the server.

Signed-off-by: Joas Schilling <coding@schilljs.com>