Increase opacity when the input field for the submit button is focused
When the input field for the room name is focused pressing enter sets
the room name without having to explicitly click on the submit button.
To hint this behaviour now the opacity of the submit button is slightly
increased when its input field is focused.

Due to the specifity of the CSS rules, the one for the focused input
needs to be defined before the other to get the full opacity of the
button when it is hovered and the input is also focused.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>