Fix empty content message in Files app when WebRTC is not supported
The EmptyContentView expects some events to be triggered in the
"localMedia" radio channel while waiting for permissions or when WebRTC
is not supported to show the proper message.

Although the "Waiting for permissions" message is never shown in the
Files app (because the call UI is shown only once the permissions are
granted or denied) all the events are triggered anyway for consistency.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>