Fix mentions being too close to the following text
Browsers merge two or more consecutive spaces, even when they are in
different elements. Due to this, when there is a trailing space inside a
mention it gets merged with the space after that mention, which causes
the mention to visually "touch" the following text despite the space
between them. To prevent that now all the spaces were removed from the
Handlebars template that renders mentions.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>