Fix opacity of submit button for guest name
The guest name now behaves like the room name regarding the opacity of
its submit button.

Both the room name and the guest name are editable text labels, so the
CSS rules were moved from the room name to the editable text label.
However, although the guest name uses the default server rules for the
input and the submit button the room name does not (they are adjusted to
better fit a header); the rules set for the editable text label are
based on the default server rules, and the room name inherits the rules
for the editable text label customizing them as needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>