Add global Content Security Policy for signaling servers
The default Content Security Policy restricts connections only to the
same domain of the page, so the domains of the web sockets used to
connect to the standalone signaling servers must be explicitly allowed.

This has to be done for Nextcloud as a whole, as Talk is embedded in
other apps by loading additional scripts in the browser, and thus it is
not possible to set the Content Security Policy in the controllers that
serve those apps.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>