Fix display name of reshare with one-to-one room not invited to
The display name of reshares with one-to-one rooms not invited to was
always set to the display name of the last participant. Now it includes
the name of both participants in the room instead.

Note that this is not needed for deleted shares, as in that case only
shares by the user, not reshares, need to be formatted.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>