Do not push browser history state again when handling history events
Whenever a room was joined a new state for that room was pushed to the
browser history. As this was always done even going back in the history
caused a new state to be pushed to the history which, in turn, caused
the previous state to be erased from the history. Due to this it was not
possible to move back and forward between the rooms in the browser
history.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>