Replace state in browser history when leaving the current room
When leaving the current room the main Talk URL ws pushed to the browser
history. However, this kept the URL to the no longer accessible room in
the browser history; in that case going back in the history failed and
caused the whole app to reload. Now the main Talk URL replaces the room
URL so going back in the history works as expected.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>