Set call related CSS classes in the call container of PublicShareAuth
For legacy reasons the "incall" CSS class, which is used to hide and
show the empty content view and the call container as needed, and the
"screensharing" CSS class, which is used to adjust the call UI when
there are shared screens, were set in the "#talk-sidebar" element of
PublicShareAuth. For consistency with the embedded Talk UI now the
classes are set in the "#call-container" instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>