Fix vertical position of submit button in public share auth page
The submit button in the chat view is positioned using its "bottom"
property to align it with the bottom of the new message form even when
there are several lines and the form grows vertically.
"publicshareauth.css" (core) sets the "top" property of submit buttons,
which causes the submit button in the chat view to be aligned to the top
of the new message form, and thus needs to be overriden.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>