Fix guest name input field and submit button in public share auth page
"guest.css" overrides the default style of inputs set in "input.scss",
which causes all inputs in the public share auth page to look like the
main password input. Due to this, the rules from "input.scss" need to be
restored for inputs in the Talk sidebar to get the proper appearance.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>