Fix creating again an existing peer
In some circumstances (for example, if the signaling server provides an
incorrect list of users) the peer that created the offer can disconnect
from the peer that received the offer without causing the peer that
received the offer to remove the peer that created the offer.

In that situation, if the peer that created the offer then tried to
connect again with the other peer the other peer created a new "Peer"
object that was used instead of the old one. However, as the old one was
not removed, its video element (which was no longer updated) was kept
visible, hiding the video element linked to the new "Peer" object and
making it look frozen.

It is not possible to ensure that the peer that received the offer will
always remove the peer that created the offer if the peer that created
the offer disconnects from the peer that received the offer, so this is
handled instead by removing the stale "Peer" objects when new offers are
received.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>