Fix reconnecting from peers with a lower session ID
In some circumstances (for example, if the signaling server provides an
incorrect list of users) the peer that received the offer can disconnect
from the peer that created the offer without causing the peer that
created the offer to disconnect from the peer that received the offer
too.

A peer only tried to connect with another peer if its session ID was
larger than the session ID of the other peer; otherwise it just waited
for the other peer to start the connection instead. This is done that
way to prevent overloading a peer joining a room.

However, in the scenario above, the peer that received the offer just
waited for the peer that created the offer to start the connection
again. As the peer that created the offer was not aware of the
disconnection it did not try to connect again, and thus they never
reconnected.

It is not possible to ensure that the peer that created the offer will
always disconnect from the peer that received the offer if the peer that
received the offer disconnects from the peer that created the offer, so
this is handled instead by making that peers with smaller session IDs
also start a connection if the peer with the larger session ID did not
in a reasonable time.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>