Fix last local audio state not remembered when the UI is reloaded
When the local audio is enabled or disabled its state is written to the
local storage in order to restore it in the next call. However, the
previous value was not properly set again when the Talk UI was reloaded;
when the MediaControlsView is created the audio is not available yet, so
calling "disableAudio()" did nothing. Moreover, as "audioDisabled" was
not set to "false" once the audio was available the audio was always
enabled in "initAudioVideoSettings".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>