Fix removal of stale peers when a screen is shared
When a screen is shared a new offer is sent, just like when a video (or
audio only) connection starts; any offer coming from the same peer has
the same id, even if the type of the offer is different and separate
Peer objects are created for each offer. Due to this now a peer is stale
only if both the id and type of the new offer matches an existing peer;
otherwise when the screen offer was received the video peer was seen as
stale and removed.

Note, however, that even if the id and type of the new offer matches an
existing peer its participant should not be marked as disconnected, as
that removes all its Peer objects. If that happens then the connection
becames unstable with the peers sending offers back and forth due to the
difference in how they are started (immediately or delayed for video
depending on the ids of both peer, immediately for screens but only for
the peer that shares), the removal of Peers when a new offer is
received, and the creation of new Peers (and, with them, new offers)
when the signaling returns the user list.

Moreover, if a stale peer is found it is going to be replaced by a fresh
peer, so there it is not really needed to perform the whole
disconnection process. Therefore, now the stale peer object is simply
ended instead (and in the case of video peers their video and speaker
are explicitly removed; this should probably be done automatically when
handling the peer end, like done for screen peers, but for the time
being an explicit call is used just like in other areas of the file).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>