Fix screen icon not shown for peer after reconnecting
When a stale peer reconnects its Peer object, and thus its video view,
is removed and created again. The video view always set the screen icon
as not available and it was later changed to available if a screen offer
was received, so if the peer sent the screen offer before the video
offer the screen icon was never set as available. Now whether the screen
is shared or not is taken into account when setting the initial state of
the video view.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>