Fix disabling video when joining a call in a room with > 5 participants
Until now, when entering a room with more than five participants the
video was disabled. However, if the local video is available it is
enabled when WebRTC starts. Due to this, if WebRTC was also started when
joining the call (that is, no previous calls were made in other rooms
since the last page reload) the video was enabled instead of disabled.
Besides that, if leaving the call with the video enabled and joining it
again the video was still enabled, as the room had not been entered
again to disable the video. To fix both issues now the video is disabled
when joining the call instead of when entering the room.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>