Make "disableVideo" consistent with "enableVideo"
When calling "disableVideo" the video was hidden even if it was not
possible to disable the video; this was done to hide the video even if
WebRTC was not started yet when "disableVideo" was called. Now that the
video is disabled when joining the call in a room with more than five
participants instead of just when entering a room with more than five
participants that special behaviour is no longer needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>