Fix stopping local media when reconnecting to a call with the MCU
When a call is (successfully) left the local media are automatically
stopped. However, when the MCU is used, if the connection fails during a
call a reconnection is triggered, which first leaves the call and then
joins it again; as the local media were stopped, when the peer joins the
call again the local media is no longer sent to the other peers. To fix
this now the local media is not stopped when leaving a call due to a
reconnection.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>