Fix not stopping local media when the room is deleted during a call
When the internal signaling server fails to return the messages for the
current room the room is left. Leaving the current room also leaves the
current call (if any), which stops the local media (and remote peers)
when successful. However, if pulling the messages of a room fails it is
very likely that leaving the call in that room fails too (for example,
if the room was deleted), which would not stop the local media (nor the
call itself, as the remote peers will still be active). Due to this, now
WebRTC is explicitly stopped when pulling signaling messages fails to
ensure that the call is properly cleaned up.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>