Fix duplicated call button when joining the same room again in Files app
The Files plugin added the call button to the DetailTabView whenever the
room was joined. However, in some cases the same room can be joined
again, which caused the button to be duplicated. For example, when the
MCU is used, if the room is left from Talk UI while the Talk tab is open
in the Files app (as it will automatically join back), or if the
connection fails during a call, which triggers a reconnection to the
call but also to the room. To handle this now the call button is not
added again when reconnecting to the active room.

Although in this case it is not really needed, for consistency now
reconnections are ignored too in the DetailFileInfoView.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>