Copy "setPageTitle" to publicshareauth.js
"Embedded" does not provide a method to change the page title, as in
general only full apps should do that. However, in the case of the Talk
UI in the public share authentication page it makes sense to change the
page title; once PublicShareAuth is based on Embedded instead of on
Application it will no longer have access to that method, so now
PublicShareAuth uses its own copy.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>