Base PublicShareAuth on Embedded instead of on Application
hen PublicShareAuth was introduced there was no Embedded class, so it
had to be based on Application. However, Application should be used only
for the main Talk UI and, as its name implies, Embedded should be used
for any embedded Talk UI.

Some minor adjustments are needed to PublicShareAuth due to the
differences between Application and Embedded: unlike in Application, the
empty content view is not created by Embedded, so there is no need to
destroy it; Embedded does not render the local video view by default, so
it needs to be explicitly rendered; there is no fullscreen button for
embedded apps, so there is no need to register its handler; and finally,
the chat view needs to be explicitly detached when the room is left, as
Embedded does not handle UI states.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>