Fix remote video visible behind avatar before being explicitly shown
When a connection is established with a peer the peer sends through the
data channel whether the video is available or not, which is used to
show and hide the avatar and the video as needed. However, before that
information arrives, the video can be already streaming; as the video
element was not hidden when added to the VideoView the video could be
playing behind the avatar until the information received in the data
channel confirms that the video is available, hiding the avatar. Now the
video is hidden as soon as it is added to the VideoView and then it is
shown only after marking it as available and enabled.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>