Fix hidding video causing audio to not be played
When a remote peer is added its remote stream is attached to a new video
element (even if the remote peer is audio only). Immediately after that
the video element is hidden, and it is kept hidden until an unmute
message for the video is received.

Sometimes (mostly reproducible with Safari) even if the remote audio was
being received no audio was played if the video element was hidden; in
those cases showing the video element (for example, by removing
"display: none" from the video element with the developer tools of the
browser) caused the audio to start playing (even if there was no video
being received).

The issue only happened when using "jQuery.hide()"; it seems that its
internal handling can somehow conflict with the audio playing in some
cases. Therefore, now the element is hidden just by adding the "hidden"
CSS class instead, which seems to cause the audio to be played without
problems.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>