Fix duplicated candidate mentions in file rooms
In regular rooms only the current participants are candidates in
mentions, but in file rooms the candidates are all the users with access
to the file, even if they are not currently participants of the room.
However, as the users with access to the file are a superset of the
current participants those participants were returned twice. Therefore,
now candidates for mentions in file rooms are got only from the users
with access to the file instead of in addition to the current
participants.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>